refactor(app): 입력받는 부분 분리
